SQL: Characters db storage type cleanup No. 3 - Note:
There is a chance that you will be getting importing errors / data truncation warnings because of ancient invalid pet names.
SELECT * FROM character_pet WHERE CHAR_LENGTH(`name`) > 21;

+++ src/server/game/Server/Protocol/Handlers/NPCHandler.cpp
@@ -562,7 +562,7 @@ void WorldSession::SendStablePetCallback(QueryResult result, uint64 guid)
 
             data << uint32(fields[1].GetUInt32());          // petnumber
             data << uint32(fields[2].GetUInt32());          // creature entry
-            data << uint32(fields[3].GetUInt32());          // level
+            data << uint32(fields[3].GetUInt16());          // level
             data << fields[4].GetString();                  // name
             data << uint8(2);                               // 1 = current, 2/3 = in stable (any from 4,5,... create problems with proper show)
 
@@ -816,7 +816,7 @@ void WorldSession::HandleStableSwapPetCallback(QueryResult result, uint32 petnum
 
     Field *fields = result->Fetch();
 
-    uint32 slot        = fields[0].GetUInt32();
+    uint32 slot        = fields[0].GetUInt8();
     uint32 creature_id = fields[1].GetUInt32();
 
     if (!creature_id)
